requirements.txt
pytest #needed for tests
pytest-qgis==2.0.* #needed for tests